Crate som_interpreter_ast

This is the interpreter for the Simple Object Machine.

Modules

block
Facilities for manipulating blocks.

class
Facilities for manipulating classes.

evaluate
Facilities for evaluating nodes and expressions.

frame
Facilities for manipulating stack frames.

hashcode
Facilities for manipulating values.

instance
Facilities for manipulating class instances.

interner
Facilities for string interning.

invokable
Facilities for invoking methods and/or primitives.

method
Facilities for manipulating class methods.

primitives
Definitions for all supported primitives.

universe
The interpreter's main data structure.

value
Facilities for manipulating values.

Macros

expect_args
Macro for checking and destructuring arguments passed to primitives.

Type Definitions

SOMRef
A strong and owning reference to an object.

SOMWeakRef
A weak reference to an object.