Struct som_interpreter_ast::universe::CoreClasses

pub struct CoreClasses {
    pub object_class: SOMRef<Class>,
    pub class_class: SOMRef<Class>,
    pub metaclass_class: SOMRef<Class>,
    pub nil_class: SOMRef<Class>,
    pub integer_class: SOMRef<Class>,
    pub double_class: SOMRef<Class>,
    pub array_class: SOMRef<Class>,
    pub method_class: SOMRef<Class>,
    pub primitive_class: SOMRef<Class>,
    pub symbol_class: SOMRef<Class>,
    pub string_class: SOMRef<Class>,
    pub system_class: SOMRef<Class>,
    pub block_class: SOMRef<Class>,
    pub block1_class: SOMRef<Class>,
    pub block2_class: SOMRef<Class>,
    pub block3_class: SOMRef<Class>,
    pub boolean_class: SOMRef<Class>,
    pub true_class: SOMRef<Class>,
    pub false_class: SOMRef<Class>,
}

The core classes of the SOM interpreter.

This struct allows to always keep a reference to important classes, even in case of modifications to global bindings by user-defined code.

Fields
object_class: SOMRef<Class>

The Object class.

class_class: SOMRef<Class>

The Class class.

metaclass_class: SOMRef<Class>

The Class class.

nil_class: SOMRef<Class>

The Nil class.

integer_class: SOMRef<Class>

The Integer class.

double_class: SOMRef<Class>

The Double class.

array_class: SOMRef<Class>

The Array class.

method_class: SOMRef<Class>

The Method class.

primitive_class: SOMRef<Class>

The Primitive class.

symbol_class: SOMRef<Class>

The Symbol class.

string_class: SOMRef<Class>

The String class.

system_class: SOMRef<Class>

The System class.

block_class: SOMRef<Class>

The Block class.

block1_class: SOMRef<Class>

The Block1 class.

block2_class: SOMRef<Class>

The Block2 class.

block3_class: SOMRef<Class>

The Block3 class.

boolean_class: SOMRef<Class>

The Boolean class.

true_class: SOMRef<Class>

The True class.

false_class: SOMRef<Class>

The False class.