som_interpreter_ast::universe

Struct Universe

pub struct Universe {
    pub interner: Interner,
    pub globals: HashMap<Interned, Value>,
    pub classpath: Vec<PathBuf>,
    pub current_frame: Gc<Frame>,
    pub core: CoreClasses<Gc<Class>>,
    pub start_time: Instant,
    pub gc_interface: &'static mut GCInterface,
}

The central data structure for the interpreter.

It represents the complete state of the interpreter, like the known class definitions, the string interner and the stack frames.

Fields
interner: Interner

The string interner for symbols.

globals: HashMap<Interned, Value>

The known global bindings.

classpath: Vec<PathBuf>

The path to search in for new classes.

current_frame: Gc<Frame>

The current frame for the operation

core: CoreClasses<Gc<Class>>

The interpreter’s core classes.

start_time: Instant

The time record of the universe’s creation.

gc_interface: &'static mut GCInterface

GC interface

Implementations

impl Universe

pub fn with_classpath(classpath: Vec<PathBuf>) -> Result<Self, Error>

Initialize the universe from the given classpath.

pub fn with_classpath_and_heap_size( classpath: Vec<PathBuf>, heap_size: usize, ) -> Result<Self, Error>

Initialize the universe from the given classpath, and given a heap size

pub fn load_class( &mut self, class_name: impl Into<String>, ) -> Result<Gc<Class>, Error>

Load a class from its name into this universe.

pub fn load_system_class( classpath: &[impl AsRef<Path>], class_name: impl Into<String>, super_class: Option<Gc<Class>>, gc_interface: &mut GCInterface, interner: &mut Interner, ) -> Result<Gc<Class>, Error>

Load a system class (with an incomplete hierarchy).

impl Universe

pub fn eval_with_frame<T: Evaluate>( &mut self, value_stack: &mut GlobalValueStack, nbr_locals: u8, nbr_args: usize, invokable: &mut T, ) -> Return

Evaluates a method or other after pushing a new frame onto the stack. The frame assumes the arguments it needs are on the global argument stack.

pub fn eval_block_with_frame( &mut self, value_stack: &mut GlobalValueStack, nbr_locals: u8, nbr_args: usize, ) -> Return

Evaluates a block after pushing a new block frame.

pub fn eval_block_with_frame_no_pop( &mut self, value_stack: &mut GlobalValueStack, nbr_locals: u8, nbr_args: usize, ) -> Return

Evaluates a block after pushing a new block frame, and the new frame doesn’t get created by popping arguments - just copying them. Implemented to help avoid an odd GC bug in the to:by:do node where a ref to a pointer, kept as a variable, ended up dangling after GC for some reason. This variable was popped off the stack and pushed back. Thanks to this function, it’s just copied from the previous one to the new one. Which may also be a speedup

pub fn intern_symbol(&mut self, symbol: &str) -> Interned

Intern a symbol.

pub fn lookup_symbol(&self, symbol: Interned) -> &str

Lookup a symbol.

pub fn has_global(&self, name: Interned) -> bool

Returns whether a global binding of the specified name exists.

pub fn lookup_global(&self, name: Interned) -> Option<Value>

Search for a global binding.

pub fn assign_global(&mut self, name: Interned, value: &Value) -> Option<()>

Assign a value to a global binding.

impl Universe

pub fn escaped_block( &mut self, value_stack: &mut GlobalValueStack, value: Value, block: Gc<Block>, ) -> Option<Return>

Call escapedBlock: on the given value, if it is defined.

pub fn does_not_understand( &mut self, value_stack: &mut GlobalValueStack, value: Value, interned_sym: Interned, args: Vec<Value>, ) -> Option<Return>

Call doesNotUnderstand: on the given value, if it is defined.

pub fn unknown_global( &mut self, value_stack: &mut GlobalValueStack, value: Value, sym: Interned, ) -> Option<Return>

Call unknownGlobal: on the given value, if it is defined.

pub fn initialize( &mut self, args: Vec<Value>, value_stack: &mut GlobalValueStack, ) -> Option<Return>

Call System>>#initialize: with the given name, if it is defined.