Crate som_interpreter_ast

This is the interpreter for the Simple Object Machine.

Modules

ast

AST specific to the AST interpreter

compiler

Generates the AST

evaluate

Facilities for evaluating nodes and expressions.

gc

To interact with the GC.

hashcode

Facilities for manipulating values.

invokable

Facilities for invoking methods and/or primitives.

nodes

Specialized AST nodes

primitives

Definitions for all supported primitives.

universe

The interpreter’s main data structure.

value

Facilities for manipulating values.

vm_objects

VM-specific objects.

Macros

get_args_from_stack

Statics

STACK_ARGS_RAW_PTR_CONST

Raw pointer to trace the global value stack as a root.

UNIVERSE_RAW_PTR_CONST

Raw pointer needed to trace GC roots. Meant to be accessed only non-mutably, hence the “CONST” in the name. TODO: actually enforce that non-mutable access.